Integrate jsonlog with the http.Server error log
Go’s http.Server may also write its own log messages relating to
things like unrecovered panics, or problems accepting and writing to
HTTP connections.

So we leverage the fact that our logger satisfies the io.Writer
interface (thanks to the Write() method that we added to it),
and set http.Server to use a regular log.Logger instance from the standard library
which writes to our own Logger as the target destination.