Correct wrong test in token usecase - Insert - success
In previous test,  we set expectation to mock repo,
and call mockRepo.Insert() directly for test.

But this is wrong, we are in usecase layer, we have to call

err := usecase.Insert(mockRepo)

with mock repo passed in, then make sure mockRepo.Insert()  is called
as expected, then make sure the return value of usecase.Insert is
what we expected.